Update to UI - NavSideBar complete - Hamburger functionality = toggle hide / show - Separate component between Nav and NavLeft
Summary: That took quite a while to get through. I had a choice between making media queries display the existing NavBar on the side or creating a completely new component. The latter choice seemed to work out better. I was relecutant to try this approach because I assumed that there would be some issue with duplicating the user query in two places. It seems to be working and no loss of functionality occured. creating the toggle funcationality was fairly simple, much simpler than anticipated.